third_party/WebKit/LayoutTests/fast/dom/Document/document-open-02.html

<!DOCTYPE HTML>
<title>document.open test for creating new window</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<body>
<script>
test(function() {
    window.open = function() {
        assert_unreached('FAIL: document.open must not invoke window.open, document.open must open a new window with only internal implementation.');
    }
    w = document.open(1, 1, 1);
    assert_equals(typeof(w), 'object', 'w should be a window object');
});
</script>
</body>
